+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## [unreleased]
 ### changed
+* 🙅‍♀️ mommy's `MOMMY_FORBIDDEN_WORDS` setting now interprets each word as a regex~ ([#103](https://github.com/FWDekker/mommy/issues/103))
 * 🚀 mommy is now 4 times as fast~ ([#106](https://github.com/FWDekker/mommy/issues/106))